+++ packages/_vue3-migration-test/src/components/index.ts
@@ -8,4 +8,5 @@ export { default as TestBaseVariableColumnGrid } from './test-base-variable-colu
 export { default as TestSlidingPanel } from './test-sliding-panel.vue';
 export { default as TestUseLayouts } from './test-use-layouts.vue';
 export { default as TestBaseSuggestions } from './suggestions/test-base-suggestions.vue';
+export { default as TestHighlight } from './test-highlight.vue';
 export { default as TestBasePanel } from './panels/test-base-panel.vue';